requirements.txt
aiohttp==3.9.3
coc.py==3.2.1
kafka-python==2.0.2
motor==3.3.2
msgspec==0.15.1
orjson==3.9.15
pendulum==3.0.0
python-dotenv==1.0.1
pymongo==4.6.1
python-snappy==0.6.1
python-dateutil==2.8.2
pytz==2023.4
redis==5.0.1
ujson==5.9.0
